% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTSCluster.R
\name{HTSCluster}
\alias{HTSCluster}
\title{Title}
\usage{
HTSCluster(y, K, conds = NULL, norm = "TMM", model = "Normal",
  transformation = "none", subset.index = NA, filterMean = FALSE,
  filterCutoff = 50, geneIDs = 1:nrow(y), parallel = TRUE,
  BPPARAM = bpparam(), ...)
}
\arguments{
\item{y}{(\emph{n} x \emph{q}) matrix of observed counts for \emph{n}
observations and \emph{q} variables}

\item{K}{Number of clusters (a single value). If \code{fixed.lambda}
contains a list of lambda values to be fixed, \code{K} corresponds to the
number of clusters in addition to those fixed.}

\item{conds}{Vector of length \emph{q} defining the condition (treatment
group) for each variable (column) in \code{y}}

\item{norm}{The type of estimator to be used to normalize for differences in
library size: (\dQuote{\code{TC}} for total count, \dQuote{\code{UQ}} for
upper quantile, \dQuote{\code{Med}} for median, \dQuote{\code{DESeq}} for
the normalization method in the DESeq package, and \dQuote{\code{TMM}} for
the TMM normalization method (Robinson and Oshlack, 2010). Can also be a
vector (of length \emph{q}) containing pre-estimated library size estimates
for each sample.}

\item{model}{Type of mixture model to use (\dQuote{\code{Poisson}} or \dQuote{\code{Normal}})}

\item{transformation}{Transformation type to be used: \dQuote{\code{voom}}, \dQuote{\code{logRPKM}}
(if \code{geneLength} is provided by user), \dQuote{\code{arcsin}}, \dQuote{\code{logit}},
\dQuote{\code{logMedianRef}}, \dQuote{\code{profile}}, \dQuote{\code{clrProfile}}, 
\dQuote{\code{log}}, \dQuote{\code{normlog}}, \dQuote{\code{logMeanRef}}, \dQuote{\code{logGMeanRef}},
\dQuote{\code{vst}}, \dQuote{\code{moderatedCPM}}, \dQuote{\code{none}}}

\item{subset.index}{Optional vector providing the indices of a subset of
genes that should be used for the co-expression analysis (i.e., row indices
of the data matrix \code{y}.}

\item{filterMean}{If \code{TRUE}, filter genes with normalized mean less than
the user-provided value in \dQuote{\code{filterCutoff}}}

\item{filterCutoff}{Value used to filter low mean normalized counts when 
\dQuote{\code{filterMean}} equals \code{TRUE}.}

\item{geneIDs}{Optional vector of gene ID's.}

\item{parallel}{If \code{FALSE}, no parallelization. If \code{TRUE}, parallel 
execution using BiocParallel (see next argument \code{BPPARAM}). A note on running 
in parallel using BiocParallel: it may be advantageous to remove large, unneeded objects 
from the current R environment before calling the function, as it is possible that R's 
internal garbage collection will copy these files while running on worker nodes.}

\item{BPPARAM}{Optional parameter object passed internally to \code{bplapply} when 
\code{parallel=TRUE}. If not specified, the parameters last registered with \code{register} 
will be used.}

\item{...}{Additional optional parameters to be passed to \code{\link{NormMixClus}} or 
\code{\link{PoisMixClus}}.}
}
\value{
Object of class \code{HTSCluster} containing the full \code{results} (itself an object
of class \code{PoisMixClus} or \code{NormMixClus}), the \code{model} used, and if
applicable, the data \code{transformation} used.
}
\description{
Title
}

