% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchContTable.R
\name{matchContTable}
\alias{matchContTable}
\title{Permute rows/columns in a contingency table comparing two data clusterings}
\usage{
matchContTable(table_1, table_2)
}
\arguments{
\item{table_1}{Partition from a data clustering}

\item{table_2}{Partition from a data clustering}
}
\value{
Permuted table
}
\description{
Blah blah blah
}

