% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormMixClus_K.R
\name{NormMixClus_K}
\alias{NormMixClus_K}
\title{Normal mixture model estimation and model selection}
\usage{
NormMixClus_K(y_profiles, K, alg.type = "EM", init.runs = 50,
  init.type = "small-em", init.iter = 20, iter = 1000, cutoff = 0.001)
}
\arguments{
\item{y_profiles}{y (\emph{n} x \emph{q}) matrix of observed profiles for \emph{n}
observations and \emph{q} variables}

\item{K}{Number of clusters (a single value).}

\item{alg.type}{Algorithm to be used for parameter estimation:
\dQuote{\code{EM}}, \dQuote{\code{CEM}}, \dQuote{\code{SEM}}}

\item{init.runs}{Number of runs to be used for the Small-EM strategy, with a default value of 50}

\item{init.type}{Type of initialization strategy to be used:
\dQuote{\code{small-em}} for the Small-EM strategy, \dQuote{\code{random}}, \dQuote{\code{CEM}},
or \dQuote{\code{SEMMax}}}

\item{init.iter}{Number of iterations to be used within each run for the
Small-EM strategry, with a default value of 20}

\item{iter}{Maximum number of iterations to be run for the chosen algorithm}

\item{cutoff}{Cutoff to declare algorithm convergence}
}
\value{
\item{probaPost }{Matrix containing the conditional
probabilities of belonging to each cluster for all observations}
\item{log.like }{Value of log likelihood} 
\item{ICL }{Value of ICL criterion} 
\item{K }{Number of clusters (equivalent to \code{K})}
}
\description{
Output of \code{NormMixClus_K} is an S3 object of class \code{NormMixClus_K}.
}
\author{
Cathy Maugis-Rabusseau, Andrea Rau
}

