% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormMixParam.R
\name{NormMixParam}
\alias{NormMixParam}
\title{Calculate the mean and variance parameters for a normal mixture model}
\usage{
NormMixParam(y_profiles, probaPost)
}
\arguments{
\item{y_profiles}{y (\emph{n} x \emph{q}) matrix of observed profiles for \emph{n}
observations and \emph{q} variables}

\item{probaPost}{Matrix containing the conditional
probabilities of belonging to each cluster for all observations}
}
\value{
\item{pi }{ ...}
\item{mu }{ ...}
\item{Sigma }{ ...}
}
\description{
Corresponds to pK_Lk_Ck model
}

