% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.NormMixClus.R
\name{plot.NormMixClus}
\alias{plot.NormMixClus}
\title{Visualize results from clustering using a Normal mixture model}
\usage{
\method{plot}{NormMixClus}(x, graphs = c("logLike", "ICL", "boxplots",
  "barplots"), threshold = 0.8, order = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{"NormMixClus"}}

\item{graphs}{Type of graph to be included in plots. May be equal to \code{c("ICL",
"BIC")} for objects of class \code{"NormMixClus"}}

\item{threshold}{Minimum threshold for conditional probabilities used in graphing}

\item{order}{If \code{TRUE}, orders boxplots by the median condiitonal probability value}

\item{...}{Additional arguments}
}
\description{
Visualize results from clustering using a Normal mixture model
}
\author{
Cathy Maugis-Rabusseau
}

