% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PoisMixClus.R
\name{plot.PoisMixClus}
\alias{plot.PoisMixClus}
\title{Visualize results from clustering using a Poisson mixture model}
\usage{
\method{plot}{PoisMixClus}(x, file.name = FALSE, graphs = c("capushe",
  "ICL", "BIC"), capushe.validation = NA, ...)
}
\arguments{
\item{x}{An object of class \code{"PoisMixClus"}}

\item{file.name}{Optional file name if plots are to be saved in a PDF file.}

\item{graphs}{Type of graph to be included in plots. May be equal to \code{c("ICL",
"BIC")} for objects of class \code{"PoisMixClus"}}

\item{capushe.validation}{Optional number of clusters to use for capushe
validation (should be less than the maximum number of clusters specified in
the \code{"PoisMixClus"} object).}

\item{...}{Additional arguments}
}
\description{
A function to visualize the clustering results obtained from a Poisson
mixture model.
}
\details{
For objects of class \code{"PoisMixClus"}, the plotting function
provides the possibility for one or all of the following visualizations:

1) ICL plot for all fitted models.

2) BIC plot for all fitted models.

5) Capushe diagnostic plots.
}
\examples{
set.seed(12345)

## Simulate data as shown in Rau et al. (2011)
## Library size setting "A", high cluster separation
## n = 2000 observations
simulate <- PoisMixSim(n = 200, libsize = "A", separation = "high")
y <- simulate$y
conds <- simulate$conditions

## Run the PMM-II model for g = 3
## "TC" library size estimate, EM algorithm
run <- PoisMixClus(y, K=c(2:3), 
                   norm = "TC", conds = conds)

## Visualization of results (not run):
## plot(run)
}
\author{
Andrea Rau
}
\references{
Rau, A., Maugis-Rabusseau, C., Martin-Magniette, M.-L., Celeux, G. (2015)
Co-expression analysis of high-throughput transcriptome sequencing data with
Poisson mixture models. Bioinformatics, doi: 10.1093/bioinformatics/btu845.

Andrea Rau, Gilles Celeux, Marie-Laure Martin-Magniette, and Cathy
Maugis-Rabusseau (2011).  Clustering high-throughput sequencing data with
Poisson mixture models. \emph{Technical report} RR-7786, Inria Saclay --
Ile-de-France.
}
\seealso{
\code{\link{PoisMixClus}}, \code{\link{PoisMixClus_K}}
}
\keyword{methods}
\keyword{models}

