% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PoisMixClus_K.R
\name{plot.PoisMixClus_K}
\alias{plot.PoisMixClus_K}
\title{Visualize results from clustering using a Poisson mixture model}
\usage{
\method{plot}{PoisMixClus_K}(x, file.name = FALSE, graphs = c("map",
  "map.bycluster", "lambda"), ...)
}
\arguments{
\item{x}{An object of class \code{"PoisMixClus_K"}}

\item{file.name}{Optional file name if plots are to be saved in a PDF file.}

\item{graphs}{Type of graph to be included in plots. May be equal to
\code{"map"}, \code{"may.bycluster"}, and/or
\code{"lambda"} for objects of class \code{"PoisMixClus_K"}}

\item{...}{Additional arguments (mainly useful for plotting)}
}
\description{
A function to visualize the clustering results obtained from a Poisson
mixture model.
}
\details{
For objects of class \code{"PoisMixClus_K"}, the plotting function provides the
possibility for the following visualizations:

1) A histogram of maximum conditional probabilities across all clusters.

2) Per-cluster boxplots of maximum conditional probabilities.

3) A global view of \eqn{\ensuremath\boldsymbol{\lambda}}{\lambda} and
\eqn{\ensuremath\boldsymbol{\pi}}{\pi} values for the selected model. When
the number of conditions <= 2, bar heights represent the value of
\eqn{\ensuremath\boldsymbol{\lambda}_k}{\lambda_k} for each cluster, and bar
width corresponds to the value of
\eqn{\ensuremath\boldsymbol{\pi_k}}{\pi_k}.
}
\examples{
set.seed(12345)

## Simulate data as shown in Rau et al. (2011)
## Library size setting "A", high cluster separation
## n = 2000 observations
simulate <- PoisMixSim(n = 200, libsize = "A", separation = "high")
y <- simulate$y
conds <- simulate$conditions

## Run the PMM-II model for g = 3
## "TC" library size estimate, EM algorithm
run <- PoisMixClus(y, K=c(2:3), 
                   norm = "TC", conds = conds)

## Visualization of results (not run):
## plot(run)
}
\author{
Andrea Rau
}
\references{
Rau, A., Maugis-Rabusseau, C., Martin-Magniette, M.-L., Celeux, G. (2015)
Co-expression analysis of high-throughput transcriptome sequencing data with
Poisson mixture models. Bioinformatics, doi: 10.1093/bioinformatics/btu845.

Andrea Rau, Gilles Celeux, Marie-Laure Martin-Magniette, and Cathy
Maugis-Rabusseau (2011).  Clustering high-throughput sequencing data with
Poisson mixture models. \emph{Technical report} RR-7786, Inria Saclay --
Ile-de-France.
}
\seealso{
\code{\link{PoisMixClus}}, \code{\link{PoisMixClus_K}}
}
\keyword{methods}
\keyword{models}

