% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoisMixClus.R
\name{PoisMixClus}
\alias{PoisMixClus}
\title{Poisson mixture model estimation and selection for a series of cluster numbers}
\usage{
PoisMixClus(y, K, conds, norm = "TMM", Kmin.init = "small-em",
  split.init = FALSE, subset.index = NA, parallel = TRUE,
  BPPARAM = bpparam(), ...)
}
\arguments{
\item{y}{(\emph{n} x \emph{q}) matrix of observed counts for \emph{n}
observations and \emph{q} variables}

\item{K}{Number of clusters (a single value or a sequence of values).}

\item{conds}{Vector of length \emph{q} defining the condition (treatment
group) for each variable (column) in \code{y}}

\item{norm}{The type of estimator to be used to normalize for differences in
library size: (\dQuote{\code{TC}} for total count, \dQuote{\code{UQ}} for
upper quantile, \dQuote{\code{Med}} for median, \dQuote{\code{DESeq}} for
the normalization method in the DESeq package, and \dQuote{\code{TMM}} for
the TMM normalization method (Robinson and Oshlack, 2010). Can also be a
vector (of length \emph{q}) containing pre-estimated library size estimates
for each sample. Note that if the user provides pre-calculated normalization
factors, the package will make use of \code{norm/sum(norm)} as normalization
factors.}

\item{Kmin.init}{Type of initialization strategy to be used for the
minimum number of clusters in a sequence:
(\dQuote{\code{small-em}} for the Small-EM strategy described in Rau et al.
(2011), and \dQuote{\code{kmeans}} for a simple \emph{K}-means
initialization)}

\item{split.init}{If \code{TRUE}, the splitting initialization strategy of
Papastamoulis et al. (2014) will be used for cluster sizes larger than the minimum number, 
the initialization strategy specified in
\code{Kmin.init.type} is used otherwise.}

\item{subset.index}{Optional vector providing the indices of a subset of
genes that should be used for the co-expression analysis (i.e., row indices
of the data matrix \code{y}.}

\item{parallel}{If \code{FALSE}, no parallelization. If \code{TRUE}, parallel 
execution using BiocParallel (see next argument \code{BPPARAM}). A note on running 
in parallel using BiocParallel: it may be advantageous to remove large, unneeded objects 
from the current R environment before calling the function, as it is possible that R's 
internal garbage collection will copy these files while running on worker nodes.}

\item{BPPARAM}{Optional parameter object passed internally to \code{bplapply} when 
\code{parallel=TRUE}. If not specified, the parameters last registered with \code{register} 
will be used.}

\item{...}{Additional optional parameters to be passed to \code{\link{PoisMixClus_K}}.}
}
\value{
\item{loglike.all }{Log likelihoods calculated for each of the fitted models}
\item{capushe }{Results of
capushe model selection, an object of class \code{"Capushe"}} 
\item{ICL.all
}{ICL values calculated for each of the fitted models}
\item{ICL.results }{Object of class
\code{PoisMixClus} giving the results from the model chosen via the ICL
criterion} 
\item{BIC.results }{Object of class \code{PoisMixClus} giving the
results from the model chosen via the BIC} 
\item{DDSE.results }{Object of
class \code{PoisMixClus} giving the results from the model chosen via the
DDSE slope heuristics criterion}
\item{Djump.results }{Object of class
\code{PoisMixClus} giving the results from the model chosen via the Djump
slope heuristics criterion} 
\item{all.results }{List of objects of class
\code{PoisMixClus} giving the results for all models}
\item{model.selection }{Type of criteria used
for model selection: \code{"DDSE"}, \code{"Djump"}, \code{"BIC"}, or
\code{"ICL"} }
}
\description{
This function implements the EM and CEM algorithms for parameter
estimation in a Poisson mixture model for clustering high throughput
sequencing observations (e.g., genes) over a sequence of numbers of clusters.
Parameters are initialized using a Small-EM strategy as described in Rau et
al. (2011) or the splitting small-EM strategy described in Papastamoulis et
al. (2014), and model selection is performed using the BIC/ICL criteria or
the slope heuristics.
}
\details{
Output of \code{PoisMixClus} is an S3 object of class \code{PoisMixClus}.

In a Poisson mixture model, the data \eqn{\mathbf{y}}{y} are assumed to come
from \emph{K} distinct subpopulations (clusters), each of which is modeled
separately; the overall population is thus a mixture of these
subpopulations. In the case of a Poisson mixture model with \emph{K}
components, the model may be written as
\deqn{f(\mathbf{y};K,\ensuremath\boldsymbol{\Psi}_K) = \prod_{i=1}^n
\sum_{k=1}^K \pi_k \prod_{j=1}^{d}\prod_{l=1}^{r_j} P(y_{ijl} ;
\ensuremath\boldsymbol{\theta}_k)}{f(y;K,\psi_K) = \prod_{i=1}^n
\sum_{k=1}^K \pi_k \prod_{j=1}^{d}\prod_{l=1}^{r_j} P(y_{ijl} ; \theta_k)}
for \eqn{i = 1, \ldots, n} observations in \eqn{l = 1, \ldots, r_j}
replicates of \eqn{j = 1, \ldots, d} conditions (treatment groups), where
\eqn{P(\cdot)} is the standard Poisson density,
\eqn{\ensuremath\boldsymbol{\Psi}_K = (\pi_1,\ldots,\pi_{K-1},
\ensuremath\boldsymbol{\theta}^\prime)}{\psi_K = (\pi_1,\ldots,\pi_{K-1},
\theta^\prime)}, \eqn{\ensuremath\boldsymbol{\theta}^\prime}{\theta^\prime}
contains all of the parameters in
\eqn{\ensuremath\boldsymbol{\theta}_1,\ldots,\ensuremath\boldsymbol{\theta}_K}{\theta_1,\ldots,\theta_K}
assumed to be distinct, and \eqn{\ensuremath\boldsymbol{\pi} =
(\pi_1,\ldots,\pi_K)^\prime}{\pi = (\pi_1,\ldots,\pi_K)^\prime} are the
mixing proportions such that \eqn{\pi_k} is in (0,1) for all \emph{k} and
\eqn{\sum_k \pi_k = 1}. We consider the following parameterization for the mean
\eqn{\ensuremath\boldsymbol{\theta}_k = (\mu_{ijlk})}{\theta = (mu_{ijlk})}.
We consider \deqn{\mu_{ijlk} = w_i s_{jl} \lambda_{jk}} where \eqn{w_i}
corresponds to the expression level of observation \emph{i},
\eqn{\ensuremath\boldsymbol{\lambda}_k =
(\lambda_{1k},\ldots,\lambda_{dk})}{\lambda_k =
(\lambda_{1k},\ldots,\lambda_{dk})} corresponds to the clustering parameters
that define the profiles of the genes in cluster \emph{k} across all
variables, and \eqn{s_{jl}} is the normalized library size (a fixed
constant) for replicate \emph{l} of condition \emph{j}.

There are two approaches to estimating the parameters of a finite mixture
model and obtaining a clustering of the data: the estimation approach (via
the EM algorithm) and the clustering approach (via the CEM algorithm).
Parameter initialization is done using a Small-EM strategy as described in
Rau et al. (2011) via the \code{\link{emInit}} function. Model selection may
be performed using the BIC or ICL criteria, or the slope heuristics.
}
\note{
Note that the \code{fixed.lambda} argument is primarily intended to be
used in the case when a single cluster is fixed to have equal clustering
parameters lambda across all conditions (i.e.,
\eqn{\lambda_{j1}=\lambda_{1}=1}); this is particularly useful when
identifying genes with non-differential expression across all conditions.. Alternatively, this
argument could be used to specify a cluster for which genes are only
expressed in a single condition (e.g., \eqn{\lambda_{11} = 1} and
\eqn{\lambda_{j1} = 0} for all \eqn{j > 1}). Other possibilities could be
considered, but note that the fixed values of lambda must satisfy the
constraint \eqn{\sum_j \lambda_{jk}s_{j.} = 1} for all \eqn{k} imposed in
the model; if this is not the case, a warning message will be printed.
}
\examples{
## Simulate data as shown in Rau et al. (2011)
## Library size setting "A", high cluster separation
## n = 200 observations

simulate <- PoisMixSim(n = 200, libsize = "A", separation = "high")
y <- simulate$y
conds <- simulate$conditions

## Run the PMM model for K = 3
## "TC" library size estimate, EM algorithm
run <- PoisMixClus_K(y, K = 3, conds = conds, norm = "TC") 
summary(run)
## Estimates of pi and lambda for the selected model
pi.est <- run$pi
lambda.est <- run$lambda
norm <- run$norm

## Calculate the conditional probability of belonging to each cluster
proba <- probaPost(y, K = 3, conds = conds, pi = pi.est, s = norm,
                   lambda = lambda.est)

## Run the PMM model for K = 3 and 4
run <- PoisMixClus(y, K = 3:4, conds = conds, norm="TC")
summary(run)
}
\author{
Andrea Rau <\url{andrea.rau@jouy.inra.fr}>
}
\references{
Anders, S. and Huber, W. (2010) Differential expression analysis
for sequence count data. \emph{Genome Biology}, \bold{11}(R106), 1-28.

Papastamoulis, P., Martin-Magniette, M.-L., and Maugis-Rabusseau, C. (2014).
On the estimation of mixtures of Poisson regression models with large number
of components. \emph{Computational Statistics and Data Analysis}: 3rd
special Issue on Advances in Mixture Models, DOI:
10.1016/j.csda.2014.07.005.

Rau, A., Maugis-Rabusseau, C., Martin-Magniette, M.-L., Celeux, G. (2015)
Co-expression analysis of high-throughput transcriptome sequencing data with
Poisson mixture models. Bioinformatics, doi: 10.1093/bioinformatics/btu845.

Rau, A., Celeux, G., Martin-Magniette, M.-L., Maugis-Rabusseau, C (2011).
Clustering high-throughput sequencing data with Poisson mixture models.
Inria Research Report 7786. Available at
\url{http://hal.inria.fr/inria-00638082}.
}
\seealso{
\code{\link{probaPost}} for the calculation of the conditional
probability of belonging to a cluster; \code{\link{PoisMixMean}} for the
calculation of the per-cluster conditional mean of each observation;
\code{\link{logLikePoisMixDiff}} for the calculation of the log likelihood
of a Poisson mixture model; \code{\link{emInit}} and \code{\link{kmeanInit}}
for the Small-EM parameter initialization strategy
}
\keyword{cluster}
\keyword{models}

