% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoisMixMean.R
\name{PoisMixMean}
\alias{PoisMixMean}
\title{Calculate the conditional per-cluster mean of each observation}
\usage{
PoisMixMean(y, K, conds, s, lambda)
}
\arguments{
\item{y}{(\emph{n} x \emph{q}) matrix of observed counts for \emph{n}
observations and \emph{q} variables}

\item{K}{Number of clusters (a single value). If \code{fixed.lambda}
contains a list of lambda values to be fixed, \code{K} corresponds to the
number of clusters in addition to those fixed.}

\item{conds}{Vector of length \emph{q} defining the condition (treatment
group) for each variable (column) in \code{y}}

\item{s}{Library size normalization factors}

\item{lambda}{(\emph{d} x \code{K}) matrix containing the estimate of
\eqn{\hat{\ensuremath\boldsymbol{\lambda}}}{\hat{\lambda}}}
}
\value{
A list of length \code{K} containing the (\emph{n} x \emph{q})
matrices of mean expression for all observations, conditioned on each of the
\code{K} clusters
}
\description{
This function is used to calculate the conditional per-cluster mean
expression for all observations. This value corresponds to
\eqn{\ensuremath\boldsymbol{\mu} = (\mu_{ijlk}) = (\hat{w}_i
\hat{\lambda}_{jk})}{\mu = (\mu_{ijlk}) = (\hat{w}_i \hat{\lambda}_{jk})}
for the PMM-I model and \eqn{\ensuremath\boldsymbol{\mu} = (\mu_{ijlk}) =
(\hat{w}_i s_{jl} \hat{\lambda}_{jk})}{\mu = (\mu_{ijlk}) = (\hat{w}_i
s_{jl}\hat{\lambda}_{jk})} for the PMM-II model.
}
\examples{
set.seed(12345)

## Simulate data as shown in Rau et al. (2011)
## Library size setting "A", high cluster separation
## n = 200 observations

simulate <- PoisMixSim(n = 200, libsize = "A", separation = "high")
y <- simulate$y
conds <- simulate$conditions
s <- colSums(y) / sum(y) 	## TC estimate of lib size

## Run the PMM-II model for K = 3
## "TC" library size estimate, EM algorithm

run <- PoisMixClus_K(y, K = 3, norm = "TC", conds = conds)
pi.est <- run$pi
lambda.est <- run$lambda

## Calculate the per-cluster mean for each observation
means <- PoisMixMean(y, K = 3, conds, s, lambda.est)

}
\author{
Andrea Rau <\url{andrea.rau@jouy.inra.fr}>
}
\references{
Rau, A., Maugis-Rabusseau, C., Martin-Magniette, M.-L., Celeux,
G. (2015) Co-expression analysis of high-throughput transcriptome sequencing
data with Poisson mixture models. Bioinformatics, doi:
10.1093/bioinformatics/btu845.

Rau, A., Celeux, G., Martin-Magniette, M.-L., Maugis-Rabusseau, C. (2011).
Clustering high-throughput sequencing data with Poisson mixture models.
Inria Research Report 7786. Available at
\url{http://hal.inria.fr/inria-00638082}.
}
\seealso{
\code{\link{PoisMixClus}} for Poisson mixture model estimation and
model selection
}
\keyword{methods}

