% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probaPost.R
\name{probaPost}
\alias{probaPost}
\title{Calculate the conditional probability of belonging to each cluster in a
Poisson mixture model}
\usage{
probaPost(y, K, conds, pi, s, lambda)
}
\arguments{
\item{y}{(\emph{n} x \emph{q}) matrix of observed counts for \emph{n}
observations and \emph{q} variables}

\item{K}{Number of clusters (a single value). If \code{fixed.lambda}
contains a list of lambda values to be fixed, \code{K} corresponds to the
number of clusters in addition to those fixed.}

\item{conds}{Vector of length \emph{q} defining the condition (treatment
group) for each variable (column) in \code{y}}

\item{pi}{Vector of length \emph{g} containing the estimate of
\eqn{\hat{\ensuremath\boldsymbol{\pi}}}{\hat{\pi}}}

\item{s}{Library size normalization factors}

\item{lambda}{(\emph{d} x \code{K}) matrix containing the estimate of
\eqn{\hat{\ensuremath\boldsymbol{\lambda}}}{\hat{\lambda}}}
}
\value{
\item{t }{(\emph{n} x \code{K}) matrix made up of the conditional
probability of each observation belonging to each of the \code{K} clusters}
}
\description{
This function computes the conditional probabilities \eqn{t_{ik}} that an
observation \emph{i} arises from the \eqn{k^{\mathrm{th}}}{kth} component
for the current value of the mixture parameters.
}
\note{
If all values of \eqn{t_{ik}} are 0 (or nearly zero), the observation
is assigned with probability one to belong to the cluster with the closest
mean (in terms of the Euclidean distance from the observation). To avoid
calculation difficulties, extreme values of \eqn{t_{ik}} are smoothed, such
that those smaller than 1e-10 or larger than 1-1e-10 are set equal to 1e-10
and 1-1e-10, respectively.
}
\examples{
## Simulate data as shown in Rau et al. (2011)
## Library size setting "A", high cluster separation
## n = 200 observations

simulate <- PoisMixSim(n = 200, libsize = "A", separation = "high")
y <- simulate$y
conds <- simulate$conditions

## Run the PMM model for K = 3
## "TC" library size estimate, EM algorithm
run <- PoisMixClus_K(y, K = 3, conds = conds, norm = "TC") 
summary(run)
## Estimates of pi and lambda for the selected model
pi.est <- run$pi
lambda.est <- run$lambda
norm <- run$norm

## Calculate the conditional probability of belonging to each cluster
proba <- probaPost(y, K = 3, conds = conds, pi = pi.est, s = norm,
                   lambda = lambda.est)

## Run the PMM model for K = 3 and 4
run <- PoisMixClus(y, K = 3:4, conds = conds, norm="TC")
summary(run)
}
\author{
Andrea Rau <\url{andrea.rau@jouy.inra.fr}>
}
\references{
Rau, A., Maugis-Rabusseau, C., Martin-Magniette, M.-L., Celeux,
G. (2015) Co-expression analysis of high-throughput transcriptome sequencing
data with Poisson mixture models. Bioinformatics, doi:
10.1093/bioinformatics/btu845.

Rau, A., Celeux, G., Martin-Magniette, M.-L., Maugis-Rabusseau, C. (2011).
Clustering high-throughput sequencing data with Poisson mixture models.
Inria Research Report 7786. Available at
\url{http://hal.inria.fr/inria-00638082}.
}
\seealso{
\code{\link{PoisMixClus}} for Poisson mixture model estimation and
model selection; \code{\link{PoisMixMean}} to calculate the conditional
per-cluster mean of each observation
}
\keyword{methods}

