% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.NormMixClus.R
\name{summary.NormMixClus}
\alias{summary.NormMixClus}
\title{Summarize results from clustering using a Poisson mixture model}
\usage{
\method{summary}{NormMixClus}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"PoisMixClus"}}

\item{...}{Additional arguments}
}
\description{
A function to summarize the clustering results obtained from a Poisson
mixture model.
}
\details{
The summary function for an object of class \code{"PoisMixClus"}
provides the number of clusters selected for the BIC, ICL, DDSE, and Djump
model selection approaches.
}
\examples{
## Simulate data as shown in Rau et al. (2011)
## Library size setting "A", high cluster separation
## n = 200 observations

simulate <- PoisMixSim(n = 200, libsize = "A", separation = "high")
y <- simulate$y
conds <- simulate$conditions

## Run the PMM model for K = 3
## "TC" library size estimate, EM algorithm
run <- PoisMixClus_K(y, K = 3, conds = conds, norm = "TC") 
summary(run)
## Estimates of pi and lambda for the selected model
pi.est <- run$pi
lambda.est <- run$lambda
norm <- run$norm

## Calculate the conditional probability of belonging to each cluster
proba <- probaPost(y, K = 3, conds = conds, pi = pi.est, s = norm,
                   lambda = lambda.est)

## Run the PMM model for K = 3 and 4
run <- PoisMixClus(y, K = 3:4, conds = conds, norm="TC")
summary(run)
}
\author{
Andrea Rau
}
\references{
Rau, A., Maugis-Rabusseau, C., Martin-Magniette, M.-L., Celeux,
G. (2015) Co-expression analysis of high-throughput transcriptome sequencing
data with Poisson mixture models. Bioinformatics, doi:
10.1093/bioinformatics/btu845.

Rau, A., Celeux, G., Martin-Magniette, M.-L., Maugis-Rabusseau, C. (2011).
Clustering high-throughput sequencing data with Poisson mixture models.
Inria Research Report 7786. Available at
\url{http://hal.inria.fr/inria-00638082}.
}
\seealso{
\code{\link{PoisMixClus}}, \code{\link{PoisMixClus_K}}
}
\keyword{methods}

