% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.NormMixClus_K.R
\name{summary.NormMixClus_K}
\alias{summary.NormMixClus_K}
\title{Summarize results from clustering using a Normal mixture model}
\usage{
\method{summary}{NormMixClus_K}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"NormMixClus"}}

\item{...}{Additional arguments}
}
\description{
A function to summarize the clustering results obtained from a Normal
mixture model.
}
\details{
The summary function for an object of class \code{"NormMixClus_K"} provides the
following summary of results:

1) Number of clusters and model selection criterion used, if applicable.

2) Number of observations across all clusters with a maximum conditional
probability greater than 90% (and corresponding percentage of total
observations) for the selected model.

3) Number of observations per cluster with a maximum conditional probability
greater than 90% (and corresponding percentage of total observations per
cluster) for the selected model.

4) \eqn{\ensuremath\boldsymbol{\mu}}{\mu} values for the selected
model.

5) \eqn{\ensuremath\boldsymbol{\pi}}{\pi} values for the selected model.
}
\examples{
## Simulate data as shown in Rau et al. (2011)
## Library size setting "A", high cluster separation
## n = 200 observations

simulate <- PoisMixSim(n = 200, libsize = "A", separation = "high")
y <- simulate$y
conds <- simulate$conditions

## Run the PMM model for K = 3
## "TC" library size estimate, EM algorithm
run <- PoisMixClus_K(y, K = 3, conds = conds, norm = "TC") 
summary(run)
## Estimates of pi and lambda for the selected model
pi.est <- run$pi
lambda.est <- run$lambda
norm <- run$norm

## Calculate the conditional probability of belonging to each cluster
proba <- probaPost(y, K = 3, conds = conds, pi = pi.est, s = norm,
                   lambda = lambda.est)

## Run the PMM model for K = 3 and 4
run <- NormMixClus(log(y+1), K = 3:4)
summary(run)
}
\author{
Andrea Rau
}
\seealso{
\code{\link{NormMixClus}}, \code{\link{NormMixClus_K}}
}
\keyword{methods}

